[ascent] cleanup
